Defend from the running the migration twice
We only migrate preferences if there are any to migrate.

+++ db/migrate/20200623140437_fix_preferences_keys.rb
@@ -1,6 +1,9 @@
 class FixPreferencesKeys < ActiveRecord::Migration
   def up
-    new_preferences = Spree::Preference.where("key like '/%'")
+    unmigrated_preferences = Spree::Preference.exists?(['key NOT LIKE ?', '/%'])
+    return unless unmigrated_preferences
+
+    new_preferences = Spree::Preference.where("key LIKE '/%'")
     new_preferences.delete_all
 
     Spree::Preference.update_all("key = '/' || key")
